Ability to use an existing instance of GA
When an angular app is embedded within an existing application, it is
typical to have a pre-existing instance of google analytics that we
want to do tracking with. This change adds the capability to do so.

To use the new feature, simply instruct the provider to use an existing
instance of GA, and it will skip trying to initialize the gaq,
specifying the site ids and loading the scripts.